refactor: test about tle subcommand was modularized
