+++ tests/cli/tle_subcommand/codes.rs
@@ -1,0 +1,66 @@
+/*
+ *  Quick Test: CLI for stress testing in competitive programming
+ *  Copyright (C) 2021 - Luis Miguel Báez
+ *  License: MIT (See the LICENSE file in the repository root directory)
+ */
+
+// TLE CODES
+pub const GEN_CPP_TLE: &str = r#"
+#include <bits/stdc++.h>
+using namespace std;
+template <typename T>
+T random(const T from, const T to) {
+    static random_device rdev;
+    static default_random_engine re(rdev());
+
+    using dist_type = typename conditional<
+        is_floating_point<T>::value,
+        uniform_real_distribution<T>,
+        uniform_int_distribution<T>
+    >::type;
+
+    dist_type uni(from, to);
+    return static_cast<T>(uni(re));
+}
+int main() {
+    #define endl '\n'
+    int n = random<int>(1e5, 2e5);
+    cout << n << endl;
+    for(int i=0;i<n;++i) cout << random<int>(1, 1e9) << " ";
+    cout << endl;
+    return 0;
+}
+"#;
+
+pub const TARGET_CPP_TLE: &str = r#"
+#include <bits/stdc++.h>
+using namespace std;
+int n;
+vector<int> A;
+int main() {
+    cin >> n;
+    A.resize(n);
+    for(auto &ref: A) cin >> ref;
+    sort(A.begin(), A.end());
+    cout << n << endl;
+    for(auto &a: A) cout << a << " ";
+    cout << endl;
+    return 0;
+}
+"#;
+
+pub const GEN_PY_TLE: &str = r#"
+from random import uniform
+n = int(uniform(int(1e5), int(2e5)))
+print(n)
+A = [int(uniform(1, int(1e9))) for _ in range(n)]
+print(*A)
+"#;
+
+pub const TARGET_PY_TLE: &str = r#"
+n = int(input())
+A = list(map(int, input().split()))
+A.sort()
+print(n)
+print(*A)
+"#;
